Add key binding for help and fix other binding
Add two new bindings: one for help ('?') and another one to delete a view ('d').  The last one is in addition to the 'delete' key, since some browser bind this to "previous page in history".

Also, the correct key code (at least on Mac's QWERTY keyboard) for the letter 'h' is 72.