dial: new view, similar to gauge
Displays a vector-based gauge on an HTML <canvas> element.

Canvas elements don't accept relative widths and heights, so each time
the view is resized, we have to recalculate the absolute size in pixels
and update the canvas' properties.

When the gauge's upper limit isn't defined by the user, it is updated to
the biggest value encountered. To make this behaviour understandable, I
made the decision to always display this maximum value alongside the
current one.

gauges.js has bunch of options (colors, animation speed, etc) which I
left to the default values. This can be easily polished up later on if
we feel the need to.