Respect inline and block elements logic
Instead of using block tags (DIV) and applying CSS to make them behave
like inline tags, use SPAN tags.

Wrap the first line of items in their
own DIV tag to ensure the following content is not impacted.  Make the
description float so that the exta attributes can be side by side with
it.