Get rid of the Gemfile.lock
The lockfile allow reproducible build, which are awesome.  However, we
also want to support multiple versions of Ruby, and a lockfile is
unfortunately not a one-size-fit-all solution in this case:

  * Older versions of bundler cannot run on recent versions of Ruby;
  * Recent versions of bundler cannot run on older versions of Ruby;
  * Some dependencies are available for some versions of Ruby but not
    all the ones we support, so a version of Ruby may be unable to run
    the bundle generated by another version of Ruby.

The lock file is only used when someone clone the repository and run
`bundle install`.  When installing with `gem install riemann-dash` (the
way recommanded in the README), the dependencies only have to be in the
bounds of what is authorized in the gemspec file, so this will only
affect a fraction of our users which are probably aware of what they are
doing.

This allows us to run the test suite on all supported versions of Ruby.